requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bitarray==3.3.1
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
dropbox==12.0.2
elastic-transport==8.17.1
elasticsearch==8.11.1
fastapi==0.115.12
h11==0.16.0
idna==3.10
packaging==25.0
pdfreader==0.1.15
pillow==11.2.1
ply==3.11
pycryptodome==3.22.0
pydantic==2.11.4
pydantic_core==2.33.2
PyPDF2==3.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
requests==2.32.3
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
stone==3.3.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
wheel==0.45.1
